1. Добавил транзакции 2. Добавил работу с балансом

+++ Nadin.Application/Transaction/Models/TransactionDto.cs
@@ -10,11 +10,11 @@ namespace Nadin.Application.Transaction.Models
 	{
 		public Guid Id { get; set; }
 
-		public long Number { get; set; }
+		public double Number { get; set; }
 
-		public long BankFee { get; set; }
+		public double BankFee { get; set; }
 
-		public long SystemFee { get; set; }
+		public double SystemFee { get; set; }
 
 		public DateTime TransactionDate { get; set; }
